Created first version of class CompanyList and added a call to sync list on login.
This class provides static methods to sync the local company list with
the server list, as well as methods to retrieve the local list.

It uses an AsyncTask to do the server request in a separate background
thread.

Right now it is fully implemented, but I haven't tested it yet.