Fixed going backwards in the funnel. Now correctly undoes changes.
Had to add scrollview back to some activities so that I could see all
the buttons (why was that removed?)

Right now the activities generally don't remember their state when you
go back to them. The exception is the logo activity, since the chosen
logo is stored in the ad manager and can be retrieved without having to
be picked again.