Fix crashing when going through funnel fast. Refactor of CompanyList to Company data
No longer crashes when going through the list fast, but the button can
still be clicked more then once to skip a stage.

CompanyList was refactored to CompanyData to better reflect what it has
evolved into. Functionality has been added to take advantage of the new
Company class, and parse the JSON into an array of Companies.

There are some outdated functions that still need to be gotten rid of
once this new stuff is implemented.